VDS-1926 Move shared code for contacts and organizations importers to parent class
